+++ crates/rooch/src/commands/abi/commands/mod.rs
@@ -1,1 +1,1 @@
-pub mod export;
+pub mod export_rooch_types;
