+++ rooch_types.yaml
@@ -1,0 +1,93 @@
+AccountAddress:
+  NEWTYPESTRUCT:
+    TUPLEARRAY:
+      CONTENT: U8
+      SIZE:
+        $serde_json::private::Number: '32'
+FunctionCall:
+  STRUCT:
+  - function_id:
+      TYPENAME: FunctionId
+  - ty_args:
+      SEQ:
+        TYPENAME: TypeTag
+  - args:
+      SEQ:
+        SEQ: U8
+FunctionId:
+  STRUCT:
+  - module_id:
+      TYPENAME: ModuleId
+  - function_name:
+      TYPENAME: Identifier
+Identifier:
+  NEWTYPESTRUCT: STR
+ModuleId:
+  STRUCT:
+  - address:
+      TYPENAME: AccountAddress
+  - name:
+      TYPENAME: Identifier
+MoveAction:
+  ENUM:
+    '0':
+      Script:
+        NEWTYPE:
+          TYPENAME: ScriptCall
+    '1':
+      Function:
+        NEWTYPE:
+          TYPENAME: FunctionCall
+    '2':
+      ModuleBundle:
+        NEWTYPE:
+          SEQ:
+            SEQ: U8
+ScriptCall:
+  STRUCT:
+  - code: BYTES
+  - ty_args:
+      SEQ:
+        TYPENAME: TypeTag
+  - args:
+      SEQ:
+        SEQ: U8
+StructTag:
+  STRUCT:
+  - address:
+      TYPENAME: AccountAddress
+  - module:
+      TYPENAME: Identifier
+  - name:
+      TYPENAME: Identifier
+  - type_args:
+      SEQ:
+        TYPENAME: TypeTag
+TypeTag:
+  ENUM:
+    '0':
+      bool: UNIT
+    '1':
+      u8: UNIT
+    '2':
+      u64: UNIT
+    '3':
+      u128: UNIT
+    '4':
+      address: UNIT
+    '5':
+      signer: UNIT
+    '6':
+      vector:
+        NEWTYPE:
+          TYPENAME: TypeTag
+    '7':
+      struct:
+        NEWTYPE:
+          TYPENAME: StructTag
+    '8':
+      u16: UNIT
+    '9':
+      u32: UNIT
+    '10':
+      u256: UNIT
